Now drawing LandmarkEdges. Also, NetBeans changed the formatting of LeastSquaresListener.
The edges are drawn according to a color code. The brighter the edge is, the greater the norm of the residual; it is capped off at an arbitrary value we should probably increase. Also, pink means the residual in r is less than 0 (i.e. the edge is trying to pull the landmark and the pose near eachother) and yellow means the edge is trying to push them far away. Of course, this color code does not reflect theta, which it probably should. I just couldn't figure out a nice, visual way to display it.

Caveats: the drawing code is not my finest moment, and I have a feeling that drawing edges makes the whole thing a whole lot slower.