SiftMatcher, a class that matches SIFT feature descriptors using a kd-tree.
The metric used in the kd-tree is unweighted squared euclidian distance in 128 dimensions feature space.
There is also a simple test in rgbdslam.test.SiftMatcherTest, which matches the features of an image with
themselves. I want to write a better test (one that actually does anything), but this was quick and easy.
I benchmarked the matching to see how quick it is. For 100 features, it takes an average of 0.001s. It is
possible that it is THAT fast because all the neartest neighbors have 0 distance. In other words, we might
be testing the best case scenario. But even if the average case is 50x slower, it is still exceptionally
fast. Yay!