RANSAC now using SIFT feature matching for initialization and consensus calculation. Moved test function, which now generates random outliers, to rgbdslam.test.RANSACTest.
I have changed RANSAC to what we decided last Friday. Now it works on a list of DescriptorMatch.Match, i.e.
a list of features that have been matched through their descriptors. At every iteration, the algorithm
selects 3 points and computes the RBT between them, just like before. The consensus computation, however,
is a lot easier because we know where the transformed feature should be: near its match in the other image.
So we just compute the euclidian distance between them, and consider that match as an inlier if the distance
is below a threshold.

The final RBT is computed using all of the inliers. The list of inliers is returned to the calling code as
an out argument (not the cleanest thing in the world).